added README and renamed and refactored trim.test.js to be more legible
